Selectively invalidate stale paths, but invalidate all when needed.
If the cfMaxPathsToInvalidate option is set, invalidate up to that
many paths. If there are more stale paths than that, invalidate
everything.

This also changes the log message terminology from CloudFront 'items' to
'paths', which is more accurate.

Changed create_invalidation to count up $items itself instead of taking
an argument.