Reorder detection so that files come first
We have multiple ways to discover paths, and
some come with a filename while others don't.
Only the first version of a detected URL is
crawled. Reorder detection so that in the event
of a duplicate, we will get the filename detected
before we get a version without the filename.