Fix path ignores not matching subdirs in URLs
Currently a pattern like /category/abc ignores
subpaths such as /category/abc/def in files
because of how we parse the filesystem. We ignore
the entire dir and implicitly ignore its children.
But when we match URL paths, we only match the
entire path and don't consider if any of its
parent "directories" are ignored. We want the
ignore behavior for URL paths to be as close to
filesystem behavior as we can get it.

Fix by changing the /?$ part of the regex to
(/.*)?$ so that each path ignore pattern matches
all subpaths as well as the exact path itself.