Add S3 opt support to import_wp2static_options
Only the core_options table had a blob_value
column, so it was necessary to know which columns
to query. We check if the option_spec has an array
type which is the only type that uses blob_value,
and we only query the blob_value column in that
case.