Switch to normal guzzle lib
@leonstafford created the
[wp2staticguzzle](https://github.com/elementor/wp2staticguzzle)
lib as a solution for
https://github.com/elementor/wp2static/issues/716
However, it has not been maintained for a while.
The S3 deployment is not working with it, so I'm
replacing it with the normal guzzle lib. This
results in S3 working again.

Q: Why does the S3 deployment not work now?

A: It worked when it was in a separate plugin
that depended on the normal guzzle lib. Now
that it has been merged into the core plugin
repo, it is implicitly using wp2staticguzzle
which fails which this error:

```
Uncaught Error: Call to undefined function GuzzleHttp\Psr7\uri_for() in /home/staticweb/wordpress/wp-content/plugins/static-deploy/vendor/aws/aws-sdk-php/src/Api/Serializer/RestSerializer.php:35
```

The potential conflicts in the referenced issue
still exist, but are less likely to encounter in
practice now because we are using a single plugin.
We can no longer cause conflicts between our own
plugins. We could still encounter conflicts with
outside plugins, but that is less common. If it
does become a problem, we can revisit vendoring
the dependencies.

In that case, we can consider using the
[mozart](https://github.com/coenjacobs/mozart)
project that Leon referenced or use leon's
[AriseByAnyOtherName](https://github.com/leonstafford/AriseByAnyOtherName)
lib if it still works with recent versions of
guzzle and PHP.

Note that the aws-sdk dependency did not cause
any conflicts that I'm aware of in
five or so years of use. The reason that guzzle
had conflicts but aws-sdk didn't is because we
had multiple plugins using guzzle, but only one
using aws-sdk.