Cache and use last job_queue ajax request data
There is flashing on page load as the admin bar
changes from "Checking Status..." to a green
"Deployed" bar.

This commit addresses that by caching the last
job_queue ajax request data and using it to
update the admin bar immediately when the page
is loaded or refreshed.

This doesn't entirely remove the flashing yet,
but does reduce it.