Don't show AdminBar button until we load data
Hide the AdminBar button until we either load
cached data or send off a request. This reduces
flashing on page load.

If there is valid cached data, the first time
the button is shown it will be rendered from that
data. If there is no cached data, then the button
will be shown with "Checking status" text until
the data is fetched.

This prevents showing "Checking status" during the
milliseconds before we parse the cached data.