Fix can't set use WPCLI for immediate queue proc
Fix that the "Process Queue Immediately" option could not
have the "Using WordPress CLI" value set.

processQueueImmediately is coded so that
'0': Disabled
'1': Use WP-Cron
'2': Use WP-CLI

But because it's set as a boolean, it can't
actually have the '2' option. Change it to a
string value with those allowed_values in order
to preserve backwards-compatibility.