Add TTL to set_blob signature
It doesn't apply to the file cache, but it will
be used in other caches. Even though blobs should
stay valid indefinitely, we don't want them to
take up space indefinitely in most cases. We set
the blob's TTL to the response's max-age since we
know it will be valid for at least that long.