Split up serving cached responses and caching
I got a little ahead of myself and kept the
cached response serving code with the code that
does the caching. The problem is that that code
runs after an entire response has been generated.
This means that even if we serve a cached
response, we still wasted time generating an
entire unused response and capturing its output.

Fix by splitting the code up into two parts.
output_cache_response attempts serve a response
and returns a success boolean. If so, we can
exit immediately which skips all the rest of the
WordPress and theme and plugin code. Otherwise,
we capture the output buffer and run the same
response caching code as before.

We also move the initial cache heuristic check to
before everything else, so we can exit even
earlier if it returns false.