Convert WordPress expire args to Memcache expires
WordPress expire args to cache functions are
always the number of seconds that the item expires in,
with 0 meaning no expiration.

Memcached expiration behaves the same for 0 and values
under 60*60*24*30 (number of seconds in 30 days). But if
the value is larger than that, it is treated as a
Unix timestamp.

If $expire is under this limit, we return it unchanged.
Otherwise, we convert it to a Unix timestamp.

See https://www.php.net/manual/en/memcached.expiration.php