Add a local cache for the script duration
In early tests, we are actually slower with the
object cache than with no cache. This is probably
because the default in-memory WP object cache is
faster than going to memcached and dealing with
network and serialization.

Address this by adding a local in-memory cache
that gets checked before memcached. If we know
the cached value already, then we just return it
from memory directly. We don't have to worry about
expiration because the local cache vanishes after
the request ends or whenever script execution
ends.

There are probably a lot of repeated calls to get
the same values multiple times within the same
request, so a simple in-memory cache actually
saves a ton of work.

This is about 10% faster than default WordPress in
basic tests.