Add hidden subcommand registration
When renaming "process_queue" to "jobs process",
we can add aliases but WP_CLI keeps the old
"process_queue" command around prominently in
the commands listing.

In order to avoid this, we have to prevent
registering the command unless it is either
going to be executed or described by "help".

This is done by adding a hook to
"find_command_to_run_pre" that adds
hidden commands at the right time. To know when
the right time is, we search argv for
"static-deploy" and then we begin matching on
the subcommand parts. If we find a complete match,
then we register the subcommand.

This lets "wp static-deploy process_queue" and
"wp help static-deploy process_queue" work without
showing process_queue in the output of
"wp help static-deploy" or "wp static-deploy".