Add CLI jobs to jobs table after completion
I would like to record CLI jobs in the table for
analysis purposes. It would be tricky to add the
jobs and then process from the table, but it's
pretty easy to add the jobs only once complete.

This adds completed jobs for the detect, crawl,
post-process, deploy, and direct-deploy commands.
full-workflow subs out to the other commands, so
its jobs get added as well.