Fix Run page error from views/ symlink
The PluginUpdateChecker checks if we are in
WP's themes or plugins directory and throws an
error if it can't determine which. It seems that
the invalid path to a bootstrap file that we pass
from run-page.php, if the views/ is inside a
symlink that points to another location, causes
PUC to throw that error. Fixing the path resolves
the issue.