Fix redirects not being detected during crawl
The crawler was seeing the ultimate 200 status and
saving the PathInfo as its own page.

The old approach of following the redirect chain
is valid, but it takes more requests and can take
significantly longer.

I changed it to not follow redirects at all, so we
only make a single request. Then I just read the
Location header to get the redirect target.