Move drop-ins to a GitHub-release only dir
Wordpress.org has a lot of rules about code
that complicate things. E.g., they don't want any
direct file access, but that's hard to avoid
when using files for the page cache.

So we move the drop-ins to a directory that
is only used for GitHub releases. Submissions
to the WordPress.org repo will not contain these
files.