Replace FilesHelperImpl with constant-based branch
Add a new constant
STATIC_DEPLOY_DIRECT_FILE_ACCESS and choose
PHP or wp file access based on that.
This works because we can use rector to define
the constant at build-time and remove dead
branches.
We can also choose to leave the branches in and
allow switching configuration at runtime. But for
WP.org release, we have to build with the constant
set to false.