Add STATIC_DEPLOY_WP_ORG_MODE constant
There are some functions that are required for
wordpress.org that are simply bad choices. Create
a constant to control the use of these bad choices
so that we can enable them for WP.org builds
but completely remove from other builds.

Functions that actually have some value should
be gated by other constants.