Add escaping for exception messages
The Plugin Check Plugin for the wordpress.org
repository requires us to escape exception
messages. This is a bad idea because we should
(and do) escape output later than that.
See https://github.com/WordPress/WordPress-Coding-Standards/issues/2374
But since it's required, I added a constant that
will enable the escaping just for the WP.org
release.