Add echoing methods for OptionRenderer
Plugin Check Plugin requires us to directly
escape all echoed output. It's not smart enough
to detect what our code is doing, so we have to
escape right with the echo. I don't want to
destroy our existing functional code for this, so
I added dedicated echoing functions that use
wp_kses to sastify the plugin checks. Since we
already know that the output is safe, the wp_kses
call is wasteful and therefore we only call it
when STATIC_DEPLOY_WP_ORG_MODE is true.