Escape/type-hint trusted values on output
The values are trusted, so escaping serves no
runtime purpose but convinces Plugin Check Plugin
that we are escaping output. In some places we
are forced to escape the same trusted value
repeatedly because PCP won't accept a variable
assignment. We must escape in the echo.

Int vals are type-hinted so we don't have to
escape them.