Restructure queries so phpcs rule passes
Restructure query code so that the WP linters
for prepared queries passes. The linters are
extremely primitive and only pass if we write
the code in a very specific way. In a lot of cases
such as working with multiple rows or values, it
is impossible to satisfy the linter. In those
cases, we add an ignore rule inline.

All the code was already prepared correctly, so
this doesn't change any behavior. It just
satisfies the linter.