Don't set max_execution_time
(breaking) Remove setting of max_execution_time.
This isn't allowed to be set on all hosts, and
was set far too broadly. If your site is too large
to crawl within your normal max_execution_time,
it's recommended to use a cron job that calls
`wp static-deploy process-queue`. This has an
  unlimited max_execution_time by default.